Removed album artist frame creator :recycle:
